download over http from remote node

+++ server/src/utils/transferData/getDownloadPath.ts
@@ -1,54 +1,31 @@
-import archiver from "archiver";
 import fs from "fs-extra";
 import fsPath from "path";
 import { TMP_FOLDER } from "../../constants";
 import { DownloadPathsInput } from "../../modules/TransferData/DownloadSessionInput";
+import { archiveFolder } from "./archiveFolder";
 
 export const getDownloadPath = async (
-  session: DownloadPathsInput[],
-  sessionId: string
+	session: DownloadPathsInput[],
+	sessionId: string
 ): Promise<{ path: string, deleteTmpFolder: boolean, err?: any }> => {
-  if (session.length === 1 && session[0].type === "file")
-    return { path: session[0].path, deleteTmpFolder: false }
+	if (session.length === 1 && session[0].type === "file")
+		return { path: session[0].path, deleteTmpFolder: false }
 
-  const folderName = `${TMP_FOLDER}/${sessionId}`;
+	const folderName = fsPath.join(TMP_FOLDER, sessionId);
 
-  fs.mkdirSync(folderName);
+	fs.mkdirSync(folderName);
 
-  for (const item of session) {
-    if (item.type === "file")
-      fs.copyFileSync(item.path, `${folderName}/${fsPath.basename(item.path)}`);
-    else fs.copySync(item.path, `${folderName}/${fsPath.basename(item.path)}`);
-  }
+	for (const item of session) {
+		if (item.type === "file")
+			fs.copyFileSync(item.path, fsPath.join(folderName, fsPath.basename(item.path)));
+		else fs.copySync(item.path, fsPath.join(folderName, fsPath.basename(item.path)));
+	}
 
-  const output = fs.createWriteStream(`${folderName}.zip`),
-    archive = archiver("zip", {
-      zlib: { level: 9 },
-    });
+	const { err, path } = await archiveFolder({ sessionId, folderName })
+	if (err) console.log(err)
+	if (!path) return { err: "could not create path", deleteTmpFolder: true, path: "" }
 
-  archive.on("warning", (err) => {
-    if (err.code === "ENOENT") console.log(err);
-    else return { err };
-  });
-  archive.on("error", (err) => {
-    return { err };
-  });
+	fs.removeSync(folderName);
 
-  archive.pipe(output);
-
-  archive.directory(folderName, `${sessionId}`);
-
-  await archive.finalize();
-
-  await new Promise((res, rej) => {
-    output.on("close", () => {
-      res("");
-    });
-
-    output.on("error", (e) => rej(e));
-  });
-
-  fs.removeSync(folderName);
-
-  return { path: `${folderName}.zip`, deleteTmpFolder: true }
+	return { path, deleteTmpFolder: true }
 };
